+++ .github/metadata.py
@@ -25,6 +25,7 @@ release_commits=6f27db47612aaee06fdf08744b09a9f5f6c2
 python_files=".github/update_mailmap.py" ".github/metadata.py" "setup.py"
 doc_files="changelog.md" "readme.md" "docs/license.md"
 is_python_project=true
+uv_requirement_params=--requirement pyproject.toml
 package_name=click-extra
 blacken_docs_params=--target-version py37 --target-version py38
 ruff_py_version=py37
@@ -366,12 +367,19 @@ class Metadata:
 
     @cached_property
     def python_files(self) -> Iterator[str]:
-        """Returns list of python files."""
+        """Returns a list of python files."""
         yield from self.glob_files("**/*.py", "!.venv/**")
 
     @cached_property
+    def requirement_files(self) -> Iterator[str]:
+        """Returns a list of requirement files supported by uv."""
+        yield from self.glob_files(
+            "**/pyproject.toml", "*requirements.txt", "requirements/*.txt"
+        )
+
+    @cached_property
     def doc_files(self) -> Iterator[str]:
-        """Returns list of doc files."""
+        """Returns a list of doc files."""
         yield from self.glob_files("**/*.{md,markdown,rst,tex}", "!.venv/**")
 
     @cached_property
@@ -380,6 +388,10 @@ class Metadata:
         if self.pyproject_path.exists() and self.pyproject_path.is_file():
             return True
         return False
+
+    @cached_property
+    def uv_requirement_params(self) -> Iterator[str]:
+        return (f"--requirement {req}" for req in self.requirement_files)
 
     @cached_property
     def pyproject(self) -> dict[str, Any] | None:
@@ -898,6 +910,7 @@ class Metadata:
             "python_files": (self.python_files, False),
             "doc_files": (self.doc_files, False),
             "is_python_project": (self.is_python_project, False),
+            "uv_requirement_params": (self.uv_requirement_params, False),
             "package_name": (self.pyproject.canonical_name, False),
             "blacken_docs_params": (self.blacken_docs_params, False),
             "ruff_py_version": (self.ruff_py_version, False),