+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## [3.5.10 (unreleased)](https://github.com/kdeldycke/workflows/compare/v3.5.9...main)
 
-> [!IMPORTANT]
+> \[!IMPORTANT\]
 > This version is not released yet and is under active development.
 
 - Replace Myst admonition in changelog by GFM alerts.