+++ changelog.md
@@ -5,6 +5,8 @@
 > \[!IMPORTANT\]
 > This version is not released yet and is under active development.
 
+- Build binaries without Poetry.
+
 ## [3.5.11 (2024-06-22)](https://github.com/kdeldycke/workflows/compare/v3.5.10...v3.5.11)
 
 - Read `pyproject.toml` without relying on Poetry.