+++ changelog.md
@@ -8,7 +8,7 @@
 - Package all utilities in a `gha_utils` CLI.
 - Remove support for Poetry-based projects. All Python projects are expected to follow [standard `pyproject.toml` conventions](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/).
 - Sort contributors in `.mailmap` files regardless of case sensitivity.
-- Force default values of workflow's inputs when triggered from other events (i.e. in non-reuseable contexts).
+- Force default values of workflow's inputs when triggered from other events (i.e. in non-reusable contexts).
 - Run all Python-based commands via `uv run` and `uv tool run`.
 - Replace `is_poetry_project` metadata by `is_python_project`.
 - Add new  and `uv_requirement_params` metadata output.