+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [4.0.1 (unreleased)](https://github.com/kdeldycke/workflows/compare/v4.0.0...main)
+
+> \[!IMPORTANT\]
+> This version is not released yet and is under active development.
+
 ## [4.0.0 (2024-06-29)](https://github.com/kdeldycke/workflows/compare/v3.5.11...v4.0.0)
 
 - Package all utilities in a `gha_utils` CLI.